Fix benchmark names
